Add mark step button during recording
Mark step adds a regioncard from the last mark step (or start of
recording) to the current time. If mark step is used, stopping recording
adds a final region card from the last mark step to the end. This allows
either the remote operator or the recorder to mark process steps as they
go.